+++ src/internal.ts
@@ -3,16 +3,16 @@
  * @see https://medium.com/visual-development/how-to-fix-nasty-circular-dependency-issues-once-and-for-all-in-javascript-typescript-a04c987cf0de
  */
 
-export * from './parser/internal';
 export * from './parser/internal/parse-result';
-export * from './parser/primary/internal/reduce';
+export * from './parser/internal/parser';
+export * from './parser/internal/reduce';
+export * from './parser/predicate';
+export * from './parser/prioritized-choice';
+export * from './parser/sequence';
 export * from './parser/primary/any-character';
 export * from './parser/primary/character-class';
 export * from './parser/primary/customizable';
 export * from './parser/primary/literal-string';
-export * from './parser/primary/predicate';
-export * from './parser/primary/prioritized-choice';
-export * from './parser/primary/sequence';
 export * from './parser/modifier/internal/converter';
 export * from './parser/modifier/internal/any-or-more';
 export * from './parser/modifier/internal/value-converter';
